Reorganized focusing on reuse in other projects
Added:
 - Example directory: A toolchain test and demo of intended use.
 - Toolchain directory: Contains linker scripts, board configurations,
   and other misc required by the toolchain but not invoking it.

This modularizes the ninja file, making it easier to be used from a
particular project. See the example directory for intended use. Once
further boards and MCUs are in use, the intent is to add
subdirectories in the toolchain directory for each type of file.

Finally this also expands the demo a small amount by having it print a
number. GCC behind the scenes can turn some printf() statements into
puts(), thus greatly shrinking the size of the binary. Including a
formatted number pulls in the full printf() for a compiled size of about
34kb in the example, too big to fit in the smallest MCUs. To work around
this, nano specs were enabled for the compiler. I didn't get them to
work in GCC 4.9.3, but they did work without complaint in 6.3.1, so
that's something to keep in mind.