Fixed gdb.sh for openocd 0.9.0 and 0.10.0
Using gdb and semihosting with openocd 0.9.0 is kind of broken, it can't
be done using the normal gdb integration ("gdb_port pipe") because the
semihosting output gets inserted into the gdb stream and mucks
everything up. 0.10.0 fixes this by introducing the command `arm
semihosting_fileio enable`.

gdb.sh now tries to distinguish between the two versions of openocd and
do a less wrong thing in each instance, though it is kind of hacky

Additionally it would be nice to somehow integrate this into the ninja
build system since it knows where the toolchain directory is and also
what $1 and $2 should be, but just straight slapping it in as a rule or
build command doesn't quite work because ninja buffers standard out
until the process quits, rendering the user unable to see the gdb
interface.