Made link rule external
It turns out the contents of ninja variables are interpreted as a single
string including whitespace instead of as a space separated list a la
Make. This means I can't pass all of the object files into the link rule
using a variable, they need to be immediately specified separately. Thus
the link rule must be invoked in the per-project ninja file because
there's no way to pass it to a global link.